@@ -1,11 +1,11 @@
 name: flutter_flipperkit
 description: Flipper (Extensible mobile app debugger) for flutter.
-version: 0.0.25
+version: 0.0.26
 author: LiJianying <lijy91@foxmail.com>
 homepage: https://github.com/leanflutter/flutter_flipperkit
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0 <2.0.0"
 
 dependencies:
