@@ -1,20 +1,19 @@
-import 'dart:async';
 import 'package:flutter/services.dart';
 import './flipper_plugin.dart';
 
 class FlipperClient {
   static const MethodChannel _channel = const MethodChannel('flutter_flipperkit'); 
 
-  static FlipperClient _flipperClient;
+  static FlipperClient? _flipperClient;
   static FlipperClient getDefault() {
     if (_flipperClient == null) {
       _flipperClient = new FlipperClient();
     }
 
-    return _flipperClient;
+    return _flipperClient!;
   }
 
-  Map<String, FlipperPlugin> _plugins;
+  Map<String, FlipperPlugin>? _plugins;
 
   FlipperClient() {
     _plugins = new Map();
@@ -24,11 +23,11 @@ class FlipperClient {
     await _channel.invokeMethod('clientAddPlugin', {
       'id': plugin.getId(),
     });
-    _plugins.putIfAbsent(plugin.getId(), () => plugin);
+    _plugins!.putIfAbsent(plugin.getId(), () => plugin);
   }
 
-  FlipperPlugin getPlugin(String id) {
-    if (_plugins.containsKey(id)) return _plugins[id];
+  FlipperPlugin? getPlugin(String id) {
+    if (_plugins!.containsKey(id)) return _plugins?[id];
     return null;
   }
 
