@@ -3,7 +3,7 @@ import './flipper_http_client.dart';
 
 class FlipperHttpOverrides extends HttpOverrides {
   @override
-  HttpClient createHttpClient(SecurityContext context) {
+  HttpClient createHttpClient(SecurityContext? context) {
     final client = super.createHttpClient(context);
 
     return new FlipperHttpClient(client);
